+++ src/huc/const.h
@@ -7,11 +7,11 @@
 #define _CONST_H
 
 void new_const (void);
-void add_const (long typ);
-long array_initializer (long typ, long id, long stor);
-long scalar_initializer (long typ, long id, long stor);
-long get_string_ptr (long typ);
-long get_raw_value (char sep);
+void add_const (intptr_t typ);
+intptr_t array_initializer (intptr_t typ, intptr_t id, intptr_t stor);
+intptr_t scalar_initializer (intptr_t typ, intptr_t id, intptr_t stor);
+intptr_t get_string_ptr (intptr_t typ);
+intptr_t get_raw_value (char sep);
 int add_buffer (char *p, char c, int is_address);
 void dump_const (void);
 char *get_const (SYMBOL *);
