+++ src/huc/expr.c
@@ -29,9 +29,9 @@ void expression (int comma)
 	expression_ex(lval, comma, NO);
 }
 
-long expression_ex (LVALUE *lval, int comma, int norval)
+intptr_t expression_ex (LVALUE *lval, int comma, int norval)
 {
-	long k;
+	intptr_t k;
 
 	do {
 		if ((k = heir1(lval, comma)) && !norval)
@@ -80,9 +80,9 @@ static int is_ptrptr (LVALUE *lval)
 	return (s && (s->ptr_order > 1 || (s->ident == ARRAY && s->ptr_order > 0)));
 }
 
-long heir1 (LVALUE *lval, int comma)
+intptr_t heir1 (LVALUE *lval, int comma)
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 	char fc;
 
@@ -151,10 +151,10 @@ long heir1 (LVALUE *lval, int comma)
 	}
 }
 
-long heir1a (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir1a (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k, lab1, lab2;
+	intptr_t k, lab1, lab2;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir1b(lval, comma);
@@ -184,10 +184,10 @@ long heir1a (LVALUE *lval, int comma)
 		return (0);
 }
 
-long heir1b (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir1b (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k, lab;
+	intptr_t k, lab;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir1c(lval, comma);
@@ -209,10 +209,10 @@ long heir1b (LVALUE *lval, int comma)
 		return (0);
 }
 
-long heir1c (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir1c (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k, lab;
+	intptr_t k, lab;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir2(lval, comma);
@@ -234,10 +234,10 @@ long heir1c (LVALUE *lval, int comma)
 		return (0);
 }
 
-long heir2 (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir2 (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir3(lval, comma);
@@ -261,10 +261,10 @@ long heir2 (LVALUE *lval, int comma)
 	}
 }
 
-long heir3 (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir3 (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir4(lval, comma);
@@ -288,10 +288,10 @@ long heir3 (LVALUE *lval, int comma)
 	}
 }
 
-long heir4 (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir4 (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir5(lval, comma);
@@ -324,10 +324,10 @@ int is_byte (LVALUE *lval)
 	return (0);
 }
 
-long heir5 (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir5 (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir6(lval, comma);
@@ -356,10 +356,10 @@ long heir5 (LVALUE *lval, int comma)
 	}
 }
 
-long heir6 (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir6 (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir7(lval, comma);
@@ -439,10 +439,10 @@ long heir6 (LVALUE *lval, int comma)
 	}
 }
 
-long heir7 (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir7 (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir8(lval, comma);
@@ -475,10 +475,10 @@ long heir7 (LVALUE *lval, int comma)
 	}
 }
 
-long heir8 (LVALUE *lval, int comma)
-/*long	lval[]; */
+intptr_t heir8 (LVALUE *lval, int comma)
+/*intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir9(lval, comma);
@@ -527,10 +527,10 @@ long heir8 (LVALUE *lval, int comma)
 	}
 }
 
-long heir9 (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir9 (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	LVALUE lval2[1] = {{0}};
 
 	k = heir10(lval, comma);
@@ -565,10 +565,10 @@ long heir9 (LVALUE *lval, int comma)
 	}
 }
 
-long heir10 (LVALUE *lval, int comma)
-/* long	lval[]; */
+intptr_t heir10 (LVALUE *lval, int comma)
+/* intptr_t	lval[]; */
 {
-	long k;
+	intptr_t k;
 	SYMBOL *ptr;
 
 	if (match("++")) {
@@ -667,7 +667,7 @@ long heir10 (LVALUE *lval, int comma)
 
 		/* global and non-array */
 		ptr = lval->symbol;
-		immed(T_SYMBOL, (long)ptr);
+		immed(T_SYMBOL, (intptr_t)ptr);
 		lval->indirect = ptr->type;
 		return (0);
 	}
@@ -714,10 +714,10 @@ long heir10 (LVALUE *lval, int comma)
 	}
 }
 
-long heir11 (LVALUE *lval, int comma)
-/*long	*lval; */
+intptr_t heir11 (LVALUE *lval, int comma)
+/*intptr_t	*lval; */
 {
-	long direct, k;
+	intptr_t direct, k;
 	SYMBOL *ptr;
 	char sname[NAMESIZE];
 
@@ -884,14 +884,14 @@ long heir11 (LVALUE *lval, int comma)
 		return (k);
 
 	if (ptr->ident == FUNCTION) {
-		immed(T_SYMBOL, (long)ptr);
+		immed(T_SYMBOL, (intptr_t)ptr);
 		return (0);
 	}
 	return (k);
 }
 
 void store (LVALUE *lval)
-/* long	*lval; */
+/* intptr_t	*lval; */
 {
 	if (lval->symbol2) {
 		/* far arrays (or special arrays) */
@@ -910,7 +910,7 @@ void store (LVALUE *lval)
 			putstk(lval->indirect);
 		else {
 			if (lval->symbol && strcmp(lval->symbol->name, "vram") == 0)
-				out_ins(I_VPUTW, (long)NULL, (long)NULL);
+				out_ins(I_VPUTW, (intptr_t)NULL, (intptr_t)NULL);
 			else if (lval->symbol)
 				putmem(lval->symbol);
 			else
@@ -920,11 +920,11 @@ void store (LVALUE *lval)
 }
 
 void rvalue (LVALUE *lval)
-/* long	*lval; */
+/* intptr_t	*lval; */
 {
 	if ((lval->symbol != 0) && (lval->indirect == 0)) {
 		if (strcmp(lval->symbol->name, "vram") == 0)
-			out_ins(I_VGETW, (long)NULL, (long)NULL);
+			out_ins(I_VGETW, (intptr_t)NULL, (intptr_t)NULL);
 		else
 			getmem(lval->symbol);
 	}
